chore: move some files around
