+++ frontend/appflowy_flutter/lib/plugins/database/application/field/type_option/select_type_option_actions.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:appflowy/plugins/database/domain/type_option_service.dart';
-import 'package:appflowy/plugins/database/grid/presentation/widgets/header/type_option/builder.dart';
+import 'package:appflowy/plugins/database/widgets/field/type_option_editor/builder.dart';
 import 'package:appflowy_backend/log.dart';
 import 'package:appflowy_backend/protobuf/flowy-database2/select_option_entities.pb.dart';
