+++ frontend/appflowy_flutter/lib/plugins/database/widgets/field/type_option_editor/date.dart
@@ -5,7 +5,7 @@ import 'package:flowy_infra_ui/flowy_infra_ui.dart';
 import 'package:flutter/material.dart';
 import 'package:protobuf/protobuf.dart';
 
-import '../../../layout/sizes.dart';
+import '../../../grid/presentation/layout/sizes.dart';
 import 'builder.dart';
 import 'date/date_time_format.dart';
