+++ frontend/appflowy_flutter/lib/plugins/database/widgets/field/type_option_editor/number.dart
@@ -10,8 +10,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:protobuf/protobuf.dart';
 
-import '../../../layout/sizes.dart';
-import '../../common/type_option_separator.dart';
+import '../../../grid/presentation/layout/sizes.dart';
+import '../../../grid/presentation/widgets/common/type_option_separator.dart';
 import 'builder.dart';
 
 class NumberTypeOptionEditorFactory implements TypeOptionEditorFactory {
