+++ frontend/appflowy_flutter/lib/plugins/database/widgets/field/type_option_editor/rich_text.dart
